# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Artwork Catalogue API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artworks:
    get:
      operationId: listArtworks
      tags: [artworks]
      summary: List all artworks
      parameters:
        - name: locationId
          in: query
          schema: { type: ["integer", "null"] }
        - name: artist
          in: query
          schema: { type: ["string", "null"] }
        - name: medium
          in: query
          schema: { type: ["string", "null"] }
        - name: onLoan
          in: query
          schema: { type: ["boolean", "null"] }
        - name: search
          in: query
          schema: { type: ["string", "null"] }
        - name: yearFrom
          in: query
          schema: { type: ["integer", "null"] }
        - name: yearTo
          in: query
          schema: { type: ["integer", "null"] }
      responses:
        "200":
          description: List of artworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artwork"
    post:
      operationId: createArtwork
      tags: [artworks]
      summary: Create a new artwork
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
...
  /artworks/{id}:
    get:
      operationId: getArtwork
      tags: [artworks]
      summary: Get a single artwork with full details
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Artwork detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtworkDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateArtwork
      tags: [artworks]
      summary: Update an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtworkUpdate"
      responses:
        "200":
          description: Updated artwork
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artwork"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteArtwork
      tags: [artworks]
      summary: Delete an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /artworks/{id}/provenance:
    get:
      operationId: listProvenance
      tags: [artworks]
      summary: List provenance records for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Provenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProvenanceRecord"
    post:
      operationId: addProvenance
      tags: [artworks]
      summary: Add a provenance record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvenanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvenanceRecord"
  /artworks/{id}/provenance/{provenanceId}:
    patch:
      operationId: updateProvenance
      tags: [artworks]
      summary: Update a provenance record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: provenanceId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvenanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvenanceRecord"
    delete:
      operationId: deleteProvenance
      tags: [artworks]
      summary: Delete a provenance record
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: provenanceId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /artworks/{id}/condition-reports:
    get:
      operationId: listConditionReports
      tags: [artworks]
      summary: List condition reports for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Condition reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConditionReport"
    post:
      operationId: addConditionReport
      tags: [artworks]
      summary: Add a condition report
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConditionReportInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConditionReport"
  /artworks/{id}/documents:
    get:
      operationId: listDocuments
      tags: [artworks]
      summary: List documents for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: addDocument
      tags: [artworks]
      summary: Add a document to an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /artworks/{id}/documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [artworks]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: documentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /artworks/{id}/pricing:
    get:
      operationId: getArtworkPricing
      tags: [artworks]
      summary: Get pricing for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Pricing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pricing"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertArtworkPricing
      tags: [artworks]
      summary: Set or update pricing for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PricingInput"
      responses:
        "200":
          description: Pricing data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pricing"
  /artworks/{id}/loans:
    get:
      operationId: listArtworkLoans
      tags: [artworks]
      summary: List loans for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
    post:
      operationId: createArtworkLoan
      tags: [artworks]
      summary: Create a loan for an artwork
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Artwork:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        artist: { type: ["string", "null"] }
        year: { type: ["integer", "null"] }
        medium: { type: ["string", "null"] }
        keywords: { type: ["string", "null"] }
        width: { type: ["number", "null"] }
        height: { type: ["number", "null"] }
        depth: { type: ["number", "null"] }
        dimensionUnit: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        locationId: { type: ["integer", "null"] }
        locationName: { type: ["string", "null"] }
        onLoan: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
...
    ArtworkInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        artist: { type: string }
        year: { type: integer }
        medium: { type: string }
        keywords: { type: string }
        width: { type: number }
        height: { type: number }
        depth: { type: number }
        dimensionUnit: { type: string }
        imageUrl: { type: string }
        notes: { type: string }
        locationId: { type: integer }
...
    ProvenanceInput:
      type: object
      properties:
        date: { type: string }
        description: { type: string }
        source: { type: string }
        price: { type: number }
        currency: { type: string }
...